Spells, Races, Classes, Monsters, and Equipment all load correctly. Download takes ~25-30seconds. maybe we can run each get in a new thread? Concurrent Lists?

+++ dnd5e-resource-browser/GenericClasses/Equipment.cs
@@ -18,7 +18,7 @@ namespace dnd5e_resource_browser
         public object damage;
         public object two_handed_damage;
         public object range;
-        public int weight;
+        public double weight;
         public APIReference[] properties;
         public string url;
     }
